elucid/process/harness/pyproject.toml
[project]
name = "elucid-harness"
version = "0.1.0"
description = "Elucid verification harness — automated pipeline verification"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
elucid = "elucid.__main__:main"
